requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-generativeai
python-dotenv
pandas
numpy
xgboost
joblib
scikit-learn
requests
PyPDF2
python-docx
beautifulsoup4
spacy
langchain
openai
langchain_community

